bare bones gui display, test cases

+++ source/src/ConvexPoint.java
@@ -1,11 +1,11 @@
 
 public class ConvexPoint {
 
-	public int x; 
-	public int y;
+	public double x; 
+	public double y;
 	public Double angle; //This will be defined once a relative angle is calculated for said point
 	
-	public ConvexPoint(int X, int Y) {
+	public ConvexPoint(double X, double Y) {
 		this.x = X; 
 		this.y = Y;
 	}
